Use a dedicated chapter class.

+++ chapterreader/chapterreader/src/main/java/de/paulwoitaschek/chapterreader/id3/ChapterMetaData.kt
@@ -2,6 +2,6 @@ package de.paulwoitaschek.chapterreader.id3
 
 internal data class ChapterMetaData(
   var id3ID: String,
-  var start: Int,
+  var start: Long,
   var title: String? = null
 )
